+++ internal/handlers/command/command.go
@@ -1,0 +1,1 @@
+package command
