+++ internal/handlers/msg_dropdatabase.go
@@ -12,43 +12,46 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shared
+package handlers
 
 import (
 	"context"
 
-	"github.com/FerretDB/FerretDB/internal/handlers/common"
 	"github.com/FerretDB/FerretDB/internal/pg"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDrop removes a collection or view from the database.
-func (h *Handler) MsgDrop(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
+// MsgDropDatabase removes the current database.
+func (h *Handler) MsgDropDatabase(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
 		return nil, lazyerrors.Error(err)
 	}
 
 	m := document.Map()
-	collection := m[document.Command()].(string)
-	db := m["$db"].(string)
+	db, ok := m["$db"].(string)
+	if !ok || db == "" {
+		return nil, lazyerrors.New("no db")
+	}
 
-	if err = h.pgPool.DropTable(ctx, db, collection); err != nil {
-		if err == pg.ErrNotExist {
-			return nil, common.NewErrorMessage(common.ErrNamespaceNotFound, "ns not found")
-		}
+	res := types.MustMakeDocument()
+	err = h.pgPool.DropSchema(ctx, db)
+	switch err {
+	case nil:
+		res.Set("dropped", db)
+	case pg.ErrNotExist:
+		// nothing
+	default:
 		return nil, lazyerrors.Error(err)
 	}
 
+	res.Set("ok", float64(1))
+
 	var reply wire.OpMsg
 	err = reply.SetSections(wire.OpMsgSection{
-		Documents: []types.Document{types.MustMakeDocument(
-			"nIndexesWas", int32(1), // TODO
-			"ns", db+"."+collection,
-			"ok", float64(1),
-		)},
+		Documents: []types.Document{res},
 	})
 	if err != nil {
 		return nil, lazyerrors.Error(err)
