+++ internal/handlers/msg_ping.go
@@ -12,45 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shared
+package handlers
 
 import (
 	"context"
-	"os"
-	"runtime"
-	"strconv"
-	"time"
-
-	"golang.org/x/text/cases"
-	"golang.org/x/text/language"
 
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgHostInfo returns an OpMsg with the host information.
-func (h *Handler) MsgHostInfo(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
-	now := time.Now().UTC()
-	hostname, err := os.Hostname()
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
+// MsgPing OpMsg containing a ping, used to test whether a server is responding to commands.
+func (h *Handler) MsgPing(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	var reply wire.OpMsg
-	err = reply.SetSections(wire.OpMsgSection{
+	err := reply.SetSections(wire.OpMsgSection{
 		Documents: []types.Document{types.MustMakeDocument(
-			"system", types.MustMakeDocument(
-				"currentTime", now,
-				"hostname", hostname,
-				"cpuAddrSize", int32(strconv.IntSize),
-				"numCores", int32(runtime.NumCPU()),
-				"cpuArch", runtime.GOARCH,
-				"numaEnabled", false,
-			),
-			"os", types.MustMakeDocument(
-				"type", cases.Title(language.English).String(runtime.GOOS),
-			),
 			"ok", float64(1),
 		)},
 	})
