+++ internal/handlers/msg_serverstatus.go
@@ -12,32 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shared
+package handlers
 
 import (
 	"context"
-	"strconv"
 
-	"github.com/FerretDB/FerretDB/internal/bson"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
-	"github.com/FerretDB/FerretDB/internal/util/version"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// For clients that check version.
-const versionValue = "5.0.42"
-
-// MsgBuildInfo returns an OpMsg with the build information.
-func (h *Handler) MsgBuildInfo(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
+// MsgServerStatus OpMsg used to get a server status.
+func (h *Handler) MsgServerStatus(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	var reply wire.OpMsg
 	err := reply.SetSections(wire.OpMsgSection{
 		Documents: []types.Document{types.MustMakeDocument(
 			"version", versionValue,
-			"gitVersion", version.Get().Commit,
-			"versionArray", types.MustNewArray(int32(5), int32(0), int32(42), int32(0)),
-			"bits", int32(strconv.IntSize),
-			"maxBsonObjectSize", int32(bson.MaxDocumentLen),
 			"ok", float64(1),
 		)},
 	})
